refactor:상태와 UI 훅으로 분리
