+++ src/pages/salary-correction/components/HistoryModal.tsx
@@ -1,14 +1,7 @@
 import Modal from "@/components/Modal";
 import StatusBadge from "./StatusBadge";
 import { Amount, Container, Label, Reason, Row } from "./HistoryModal.styled";
-interface HistoryModalProps {
-	isOpen: boolean;
-	onClose: () => void;
-	status: string;
-	correctionType: string;
-	reason: string;
-	history: number | string;
-}
+import type { HistoryModalProps } from "../types/historyModal";
 
 const HistoryModal = ({
 	isOpen,
