+++ src/pages/salary-correction/types/historyModal.ts
@@ -1,0 +1,8 @@
+export interface HistoryModalProps {
+	isOpen: boolean;
+	onClose: () => void;
+	status: string;
+	correctionType: string;
+	reason: string;
+	history: number | string;
+}
